Extracting header to other widget
